Don't wrap deadlock exceptions to unknown storage ones.
From time to time, we may see deadlock exceptions wrapped to
unknown storage ones in logs. In general, they are fine because
provide clear understanding that some entities could not be altered
due to deadlocks. But the problem is that we wrap it to uknownstorage
exceptions that don't allow us properly (precisely) distinct such
types of errors when we try to use them in retry decorators.
We may do several retries when get deadlock exceptions but must not do
this when get a general unknown storage exception.