Treat common DatabaseError when AbstractDialectCommand.execute.
This MR unifies processing DatabaseError exceptions based on errno,
but not their precise types. It means that any exception with errno
1213 or 1062 is converted to Deadlock or Conflict exception accordingly.

Why it is needed? It was discovered that sometimes InternalError(1213)
raises when AbstractDialectCommand.execute executes. As this type of
error wasn't in the list of treating exceptions, we got UnknownStorageException,
but not Deadlock exception. Because of this, the process of retry on
error was not started.